[refactor] Auth 서비스의 HTTP 의존성 제거
