+++ mopl-api/src/main/java/com/mopl/domain/auth/dto/TokenResultDto.java
@@ -1,0 +1,7 @@
+package com.mopl.domain.auth.dto;
+
+public record TokenResultDto(
+        JwtDto jwtDto,
+        String refreshToken
+) {
+}
